Introduces the use of local storage into the application
This commit is an attempt to reduce the number of RPC calls to the
server. By caching the user's profile data locally using the storage
space introduced with HTML 5, the number of RPC calls to retrieve
the profile data is reduced.

The code has been introduced in two screens, the ContestDetailsScreen
as well as the ProfileScreen.