Bug fix for localization breaking listbox values.
BUG:
Previously, the value of listboxes were retrieved using getItemText.
This broke down when localization is applied as getItemText will
retrieve the localized values instead of the values in English.

FIX:
I introduced a LocalUtil class that contains a set of switch
statements for each ListBox that needs to be localized.