Squashed commit containing a whole lot of stuff.
This commit is a squashed combination of about 120~ or so commits
rushed out over the past week as the deadline drew closer. I did
not want the mess of commits to dirty up the centralized repository
especially since the commit messages were a non-descriptive mess
so I decided to squash it all into one little commit.