Modified registration counter to use SQL calls instead of Infinispan.
The application requires a way to keep track of how many people
has registered to accurately assign timeslots. Previously, a counter
was stored in the server's cache using Infinispan. However, I overlooked
the fact that this method will not be persistant across reboot and
will require some sort of backing file. In light of this, I changed
the code to peform a COUNT(*) query to the database instead.

I also added a create_date and last_modified timestamp to the
contestant schema.